Big refactor
First time I have looked at this for a 7 months. Can tell I was a bit out of the Laravel swing when I was making this. Have begun refactoring, but can see I'm in the middle of some user invitations here, so going to commit to this branch and replan some of this site